chore: move tasks components in subfolder
