+++ src/components/main/tasks/AddTask.tsx
@@ -4,8 +4,8 @@ import { useTranslation } from 'react-i18next';
 import { Container } from '@mui/material';
 import TextField from '@mui/material/TextField';
 
-import { TaskType } from '../../config/appDataTypes';
-import { DEFAULT_TASK, DEFAULT_TASK_DATA } from '../../config/constants';
+import { TaskType } from '../../../config/appDataTypes';
+import { DEFAULT_TASK, DEFAULT_TASK_DATA } from '../../../config/constants';
 
 type AddTaskProps = {
   addTask: (task: TaskType) => void;
